+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+unreleased
+++++++++++++++++++++
+
+* Add ftp publishing url to create webapp and show webapp
+
 0.1.2 (2017-04-03)
 ++++++++++++++++++++